corresponding sdk changes
