+++ sdk/v2/authz/role_assignments.go
@@ -2,12 +2,10 @@ package authz
 
 import (
 	"context"
-	"encoding/json"
 	"net/http"
 
 	rm "github.com/brigadecore/brigade/sdk/v2/internal/restmachinery"
 	libAuthz "github.com/brigadecore/brigade/sdk/v2/lib/authz"
-	"github.com/brigadecore/brigade/sdk/v2/meta"
 	"github.com/brigadecore/brigade/sdk/v2/restmachinery"
 )
 
@@ -19,42 +17,15 @@ const (
 	PrincipalTypeUser libAuthz.PrincipalType = "USER"
 )
 
-// RoleAssignment represents the assignment of a Role to a principal such as a
-// User or ServiceAccount.
-type RoleAssignment struct {
-	// Role assigns a Role to the specified principal.
-	Role libAuthz.Role `json:"role"`
-	// Principal specifies the principal to whom the Role is assigned.
-	Principal libAuthz.PrincipalReference `json:"principal"`
-}
-
-// MarshalJSON amends RoleAssignment instances with type metadata so that
-// clients do not need to be concerned with the tedium of doing so.
-func (r RoleAssignment) MarshalJSON() ([]byte, error) {
-	type Alias RoleAssignment
-	return json.Marshal(
-		struct {
-			meta.TypeMeta `json:",inline"`
-			Alias         `json:",inline"`
-		}{
-			TypeMeta: meta.TypeMeta{
-				APIVersion: meta.APIVersion,
-				Kind:       "RoleAssignment",
-			},
-			Alias: (Alias)(r),
-		},
-	)
-}
-
 // RoleAssignmentsClient is the specialized client for managing RoleAssignments
 // with the Brigade API.
 type RoleAssignmentsClient interface {
 	// Grant grants the system-level Role specified by the RoleAssignment to the
 	// principal also specified by the RoleAssignment.
-	Grant(context.Context, RoleAssignment) error
+	Grant(context.Context, libAuthz.RoleAssignment) error
 	// Revoke revokes the system-level Role specified by the RoleAssignment for
 	// the principal also specified by the RoleAssignment.
-	Revoke(context.Context, RoleAssignment) error
+	Revoke(context.Context, libAuthz.RoleAssignment) error
 }
 
 type roleAssignmentsClient struct {
@@ -75,7 +46,7 @@ func NewRoleAssignmentsClient(
 
 func (r *roleAssignmentsClient) Grant(
 	ctx context.Context,
-	roleAssignment RoleAssignment,
+	roleAssignment libAuthz.RoleAssignment,
 ) error {
 	return r.ExecuteRequest(
 		ctx,
@@ -90,7 +61,7 @@ func (r *roleAssignmentsClient) Grant(
 
 func (r *roleAssignmentsClient) Revoke(
 	ctx context.Context,
-	roleAssignment RoleAssignment,
+	roleAssignment libAuthz.RoleAssignment,
 ) error {
 	queryParams := map[string]string{
 		"roleType":      string(roleAssignment.Role.Type),
